+++ src/main/java/com/seahere/backend/auth/controller/AuthController.java
@@ -4,7 +4,7 @@ import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.seahere.backend.auth.jwt.exception.ValidateTokenException;
 import com.seahere.backend.auth.jwt.util.CookieUtil;
-import com.seahere.backend.company.response.CompanyResponse;
+import com.seahere.backend.company.controller.response.CompanyResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
