+++ src/main/java/com/seahere/backend/company/controller/CompanyController.java
@@ -1,12 +1,11 @@
 package com.seahere.backend.company.controller;
 
-import com.seahere.backend.company.request.CompanyCreateReq;
-import com.seahere.backend.company.request.CompanySearch;
-import com.seahere.backend.company.response.CompanyResponse;
+import com.seahere.backend.company.controller.request.CompanyCreateReq;
+import com.seahere.backend.company.controller.request.CompanySearch;
+import com.seahere.backend.company.controller.response.CompanyResponse;
 import com.seahere.backend.company.service.CompanyService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.coyote.Response;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
