+++ src/main/java/com/seahere/backend/company/service/CompanyServiceImpl.java
@@ -4,9 +4,9 @@ import com.seahere.backend.company.entity.CompanyEntity;
 import com.seahere.backend.company.exception.CompanyNotFound;
 import com.seahere.backend.company.repository.CompanyRepository;
 import com.seahere.backend.company.repository.CompanySelectRepository;
-import com.seahere.backend.company.request.CompanyCreateReq;
-import com.seahere.backend.company.request.CompanySearch;
-import com.seahere.backend.company.response.CompanyResponse;
+import com.seahere.backend.company.controller.request.CompanyCreateReq;
+import com.seahere.backend.company.controller.request.CompanySearch;
+import com.seahere.backend.company.controller.response.CompanyResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
